Move config to outside library
