+++ simulation/__init__.py
@@ -1,6 +1,8 @@
 # Copyright 2022-, Semiotic AI, Inc.
 # SPDX-License-Identifier: Apache-2.0
 
+from typing import Any
+
 import numpy as np
 import sacred
 from simulation.environment import Environment
@@ -8,45 +10,17 @@ from simulation.environment import Environment
 simulation_ingredient = sacred.Ingredient("simulation")
 
 
-@simulation_ingredient.config
-def config():
-    nproducts = 1
-    isa = {"kind": "softmax", "source": "consumer", "to": "indexer"}
-    entities = [
-        {
-            "kind": "entity",
-            "count": 1,
-            "group": "consumer",
-            "state": {
-                "kind": "budget",
-                "low": 0,
-                "high": 1,
-                "initial": 0.5 * np.ones(nproducts),
-                "traffic": np.ones(nproducts),
-            },
-        },
-        {
-            "kind": "agent",
-            "count": 1,
-            "group": "indexer",
-            "state": {
-                "kind": "price",
-                "low": np.zeros(nproducts),
-                "high": 3 * np.ones(nproducts),
-                "initial": np.ones(nproducts),
-            },
-            "action": {
-                "kind": "pricemultiplier",
-                "low": np.zeros(nproducts),
-                "high": 3 * np.ones(nproducts),
-                "shape": (nproducts,),
-                "baseprice": 2 * np.ones(nproducts),
-            },
-            "reward": [
-                {
-                    "kind": "traffic",
-                    "multiplier": 1,
-                }
-            ],
-        },
-    ]
+@simulation_ingredient.capture
+def environment(
+    isa: dict[str, Any], entities: list[dict[str, Any]], **kwargs
+) -> Environment:
+    """Construct an environment from the simulation config.
+
+    Arguments:
+        isa (dict[str, Any]): The config for the ISA.
+        entities (list[dict[str, Any]]): The configs for each group of entities.
+
+    Returns:
+        Environment: An instantiated simulation environment.
+    """
+    return Environment(isa=isa, entities=entities)
