+++ tests/fixture.py
@@ -42,3 +42,53 @@ def consumerconfig():
             "traffic": np.ones(3),
         },
     }
+
+
+@pytest.fixture
+def simulationconfig():
+    nproducts = 1
+    return {
+        "isa": {"kind": "softmax", "source": "consumer", "to": "indexer"},
+        "entities": [
+            {
+                "kind": "entity",
+                "count": 1,
+                "group": "consumer",
+                "state": {
+                    "kind": "budget",
+                    "low": 0,
+                    "high": 1,
+                    "initial": 0.5 * np.ones(nproducts),
+                    "traffic": np.ones(nproducts),
+                },
+            },
+            {
+                "kind": "agent",
+                "count": 2,
+                "group": "indexer",
+                "state": {
+                    "kind": "price",
+                    "low": np.zeros(nproducts),
+                    "high": 3 * np.ones(nproducts),
+                    "initial": np.ones(nproducts),
+                },
+                "action": {
+                    "kind": "price",
+                    "low": np.zeros(nproducts),
+                    "high": 3 * np.ones(nproducts),
+                    "shape": (nproducts,),
+                },
+                "reward": [
+                    {
+                        "kind": "traffic",
+                        "multiplier": 1,
+                    }
+                ],
+                "observation": [
+                    {
+                        "kind": "bandit",
+                    }
+                ],
+            },
+        ],
+    }
